requirements.txt
# Core NLP (for local NER only)
transformers>=4.38.0
torch>=2.0.0

# Gemini API
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Web Interface
streamlit>=1.31.0
